+++ atdts/test/ts-tests/test_atdts.ts
@@ -57,6 +57,18 @@ function test_everything() {
     nullables: [13, 71],
     untyped_things: [{}, [["hello"]], 123],
     foo: null,
+    parametrized_record: {
+      field_a: 42,
+      field_b: [
+        9.9,
+        8.8
+      ]
+    },
+    parametrized_tuple: [
+      { kind: 'WOW' },
+      { kind: 'WOW' },
+      100
+    ],
   }
   const a_str = JSON.stringify(API.writeRoot(a_obj), null, 2)
   save('a_str', a_str)
@@ -159,7 +171,19 @@ ${a_str}`
     ],
     123
   ],
-  "foo": null
+  "foo": null,
+  "parametrized_record": {
+    "field_a": 42,
+    "field_b": [
+      9.9,
+      8.8
+    ]
+  },
+  "parametrized_tuple": [
+    "wow",
+    "wow",
+    100
+  ]
 }`
   save('b_str', b_str)
   const b_obj = API.readRoot(JSON.parse(a_str))