+++ CHANGES.md
@@ -1,5 +1,6 @@
 * atdpy: Support parametrized type definitions (#303)
 * atdpy: Support options approximately by treating them as nullables (#320)
+* atdts: Support parametrized type definitions (#303)
 
 2.10.0 (2022-08-09)
 -------------------